AgentSignal -> TimerSignal, move to timer/mod.rs
