+++ src/timer/mod.rs
@@ -4,6 +4,27 @@
 // https://www.apache.org/licenses/LICENSE-2.0>. This file may not be copied, modified, or distributed
 // except according to those terms.
 
+
+/// A signal sent to the timer.
+///
+/// Either schedules another wake-up, or asks
+/// the timer thread to terminate.
+#[derive(Debug, Clone, Copy)]
+pub enum TimerSignal {
+    // Thread termination was requested.
+    Terminate,
+    // When to take the next snapshot using the `Backend`.
+    NextSnapshot(u64),
+}
+
+impl std::fmt::Display for TimerSignal {
+    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
+        match self {
+            Self::Terminate => write!(f, "Terminate"),
+            Self::NextSnapshot(when) => write!(f, "NextSnapshot({})", when),
+        }
+    }
+}
 // Possibly: ios, netbsd, openbsd, freebsd
 #[cfg(target_os = "macos")] pub mod kqueue;
 #[cfg(target_os = "macos")] pub use kqueue::Timer;
